refactor: split app logic into modules
